refactor: 단위테스트 public 접근제어자 제거
