+++ src/test/java/kitchenpos/menu/validator/MenuValidatorTest.java
@@ -15,7 +15,7 @@ import kitchenpos.menu.testfixture.MenuTestFixture;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.domain.MenuGroupTestFixture;
+import kitchenpos.menugroup.testfixture.MenuGroupTestFixture;
 import kitchenpos.menugroup.repository.MenuGroupRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.testfixture.ProductTestFixture;
@@ -29,7 +29,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-public class MenuValidatorTest {
+class MenuValidatorTest {
 
     @Mock
     private MenuGroupRepository menuGroupRepository;
