+++ src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -20,7 +20,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @DisplayName("메뉴 그룹 관련 비즈니스 테스트")
 @ExtendWith(MockitoExtension.class)
-public class MenuGroupServiceTest {
+class MenuGroupServiceTest {
 
     @Mock
     private MenuGroupRepository menuGroupRepository;
